feat: Add an URL normalization component
The bar takes either a full URL (mostly from mobile apps) or a domain without 
protocol (mostly in case of web apps). This component takes care of normalizing all 
cases to a full url object, in accordance to the `secure` optional parameter.

The purpose is to replace all other URL code in cozy-bar.