fix: Avoid cozy-ui stylesheet conflict with the apps
Bundle cozy-ui instead of using the transpiled version, to avoid importing two cozy-ui stylesheets version (one in the bar and one in the app).

BREAKING CHANGE: previously, the bar imported the whole cozy-ui transpiled stylesheet. Now, only the styles of imported react component are bundled. This can lead to visual problems if an app relied on that (consciously or not). You must check your whole app when you update the cozy-bar.