fix: Change iframe visibility
We can have MUI component loaded inside an iframe. It works
pretty well on Chromium and Safari, but not on Firefox
'cauz of this bug: https://bugzilla.mozilla.org/show_bug.cgi?id=548397

Our CozyMUITheme is based on getComputedStyle, but since the iframe
is hidden, firefox crash during its creation.

The workaround is to not use display:none, but instead visibility
hidden with height & width = 0

(Our bug case was the searchbar on Drive)