fix: Button alignment via upgrade UI
Chrome introduced a new default property for button alignment and
we have fixed this in cozy-ui.

This commit is done on a hotfix branch since we need to update the bar
in the stack and the master version of the bar does not currently work
well from the stack : upgrading the bar to the latest versions of cozy-ui and
cozy-client has been done on the master branch. It works well when
the cozy-bar code is bundled with the app (for example in mobile contexts)
but crashes when the code of the bar is inserted by the stack, due to
conflicting versions of React (the bar bundling its own version of React).

So the button alignment fix was backported to the 22.3.1 version
of cozy-ui (latest version of cozy-ui for a working cozy-bar),
resulting in 22.3.2. cozy-ui@22.3.2 is installed here.