refactor: Remove aliases to React and ReactDOM
Those aliases were used so that the bar used the React from the app
if it was provided. Now, in the next version of the bar, either the
app chooses to use the bar+react, or it uses only the bar and the bar
expects to find React and ReactDOM in the environment.