fix: Remove extra padding on mobile in flagship
We use the ConfirmModal to display cozy-bar AppsMenu and UserMenu. It is a good basis because it handles opening/closing/close button/zIndex/etc but it also comes with some padding that we do not want in these modals that are special. So here we remove an extra padding that is added only in flagship.